Use OptionEntry instead of argparse
OptionEntry is designed in a way that, it can handle GApplication command line flags in a multi process environment correctly.
With argparse, running `updater.py --help` while another updater instance is running will make the first instance print the help message and terminate. Both is not intended but can barely be worked around.
Using GLib code is able to fix this.